Allow overriding case titles in case pages
Case pages inherit from a customised `case/_layout.njk` file which
itself inherits from `partials/layout.njk`. However, they currently do a
hard override on the `pageTitle` block which makes it not possible to
manually set a different `pageTitle` from inside an individual page like
it's possible in the rest of the app.

This changes the override to be a `default()` instead.

Prerequisite to https://trello.com/c/5BACPms1.