Fix summary card heading width (#270)
In a previous PR we added `width: 100%` to the h2 in the summary card
component in order to allow nesting a right aligned tag inside the h2:
https://github.com/ministryofjustice/hmpps-manage-supervisions/pull/266

However, this breaks on the summary cards which contain an `actions`
section, which requires that the h2 be left with width auto.  Otherwise,
the actions block is squished by the h2.

To resolve this, we can make the h2 have `width: 100%` if it has no
siblings.