Display error summaries when failing validation
During the arrange appointment steps, if a user fails validation, we
should display a standard `govukErrorSummary` component at the top of
the page, which links to each error message in turn. We should also
auto-focus this area and prefix the title of the page with `Error: `.

The auto-focusing is automatically done by importing the GOVUK Design
System JS, so this PR implements the rest by adding the component to the
`wizard-form.njk` shared layout, and implementing a filter to generate
the appropriate `href`s.

A bug with this is that it doesn't work for every single error, but
fixing those will come in a follow-up PR. This already does enough.