Pull h1 and nav out of main element for cases
The skip link takes users to the `<main>` on the page, but the `<main>`
is followed by a long navigation element which the user has to go
through before they can reach the actual main content. Most websites
place the navigation before the skip link anchor point (including e.g.
DAC's website).

This moves the `<h1>` and `<nav>` in the `<main>` on case pages to just
before the `<main>`.

Visually, the pages should all look identical to before.