Use abbreviation elements for all acronyms (#315)
* Add nunjucks abbr filter

* Use abbreviation elements for all acronyms

These can be helpful for people with cognitive disabilities or who are
simply unfamiliar with the definitions.

Got the definitions from https://github.com/ministryofjustice/acronyms.

* Add method to Cypress tests to find dd cell by abbreviation

Abbreviations are a nested element in `dd` elements, so we can't run
`.siblings` directly on them. We need to run it on their `.parent`.