Revert the way current/previous concerns are displayed
Previously in https://github.com/ministryofjustice/hmpps-manage-supervisions/pull/385/commits/070950ef4f6bc450d62ce5be475073affb0324ae
we changed the logic for displaying current and previous concerns.

According to http://ms-design-history.herokuapp.com/risk-tab/#incorporating-risk-of-harm-to-themselves
this behaviour was intentional and should have been left as is, because
Delius captures information for concerns in pairs (suicide is paired
with self-harm for example) and the API exposes duplicated data.